upgrade opentracing-javascript to latest upstream (#1)
* simplistic error logging

Among other things, syntax errors will not result in a stacktrace.

* remove nascent flush support

This is removed from the interface in a future upstream version, and
the current implementation doesn't do anything interesting anyway.

* update to opentracing-javascript 0.12.1

See upstream changelog for full detains.  Highlights that caused churn:
 * Must go through `globalTracer()` to get the singleton.
 * I am unclear on how `childOf` was being handled before, but it
   needs to be explicit now.
 * The setting of operation name is also explicit.

NOTE: There is a more recent tag on github, but 0.12.1 is the most
recent on npm.